succesful run with worker memory set to: 21 GB
